+++ lib/graphql.rb
@@ -1,11 +1,13 @@
 require "json"
-require "parslet"
 require "set"
 require "singleton"
 require "forwardable"
 
 module GraphQL
-  class ParseError < StandardError
+  class Error < StandardError
+  end
+
+  class ParseError < Error
     attr_reader :line, :col, :query
     def initialize(message, line, col, query)
       super(message)
@@ -16,26 +18,20 @@ module GraphQL
   end
 
   # Turn a query string into an AST
-  # @param string [String] a GraphQL query string
-  # @param as [Symbol] If you want to use this to parse some _piece_ of a document, pass the rule name (from {GraphQL::Parser})
+  # @param [String] a GraphQL query string
   # @return [GraphQL::Language::Nodes::Document]
-  def self.parse(string, as: nil)
-    parser = as ? GraphQL::PARSER.send(as) : GraphQL::PARSER
-    tree = parser.parse(string)
-    document = GraphQL::TRANSFORM.apply(tree)
-    if !document.is_a?(GraphQL::Language::Nodes::Document)
-      raise("Parse failed! Sorry, somehow we failed to turn this string into a document. Please report this bug!")
-    end
-    document
-  rescue Parslet::ParseFailed => error
-    line, col = error.cause.source.line_and_column(error.cause.pos)
-    raise GraphQL::ParseError.new(error.message, line, col, string)
+  def self.parse(query_string)
+    parse_with_racc(query_string)
+  end
+
+  def self.parse_with_racc(string)
+    GraphQL::Language::Parser.parse(string)
   end
 end
 
 # Order matters for these:
 
-require 'graphql/definition_helpers'
+require 'graphql/define'
 require 'graphql/base_type'
 require 'graphql/object_type'
 
@@ -56,16 +52,17 @@ require 'graphql/float_type'
 require 'graphql/id_type'
 require 'graphql/int_type'
 require 'graphql/string_type'
+require 'graphql/directive'
 
 require 'graphql/introspection'
 require 'graphql/language'
-require 'graphql/directive'
 require 'graphql/schema'
 require 'graphql/schema/printer'
 
 # Order does not matter for these:
 
 require 'graphql/execution_error'
+require 'graphql/invalid_null_error'
 require 'graphql/query'
 require 'graphql/repl'
 require 'graphql/static_validation'