+++ lib/graphql/directive.rb
@@ -3,26 +3,31 @@
 # {Directive} & {DirectiveChain} implement `@skip` and `@include` with
 # minimal impact on query execution.
 class GraphQL::Directive
-  include GraphQL::DefinitionHelpers::DefinedByConfig
-  attr_accessor :on, :arguments, :name, :description
-  defined_by_config :on, :arguments, :name, :description, :resolve
+  include GraphQL::Define::InstanceDefinable
+  accepts_definitions :locations, :name, :description, :include_proc, argument: GraphQL::Define::AssignArgument
+
+  attr_accessor :locations, :arguments, :name, :description
 
   LOCATIONS = [
-    ON_OPERATION =  :on_operation?,
-    ON_FRAGMENT =   :on_fragment?,
-    ON_FIELD =      :on_field?,
+    QUERY =               :QUERY,
+    MUTATION =            :MUTATION,
+    SUBSCRIPTION =        :SUBSCRIPTION,
+    FIELD =               :FIELD,
+    FRAGMENT_DEFINITION = :FRAGMENT_DEFINITION,
+    FRAGMENT_SPREAD =     :FRAGMENT_SPREAD,
+    INLINE_FRAGMENT =     :INLINE_FRAGMENT,
   ]
 
-  LOCATIONS.each do |location|
-    define_method(location) { self.on.include?(location) }
+  def initialize
+    @arguments = {}
   end
 
-  def resolve(arguments, proc)
-    @resolve_proc.call(arguments, proc)
+  def include?(arguments)
+    @include_proc.call(arguments)
   end
 
-  def resolve=(resolve_proc)
-    @resolve_proc = resolve_proc
+  def include_proc=(include_proc)
+    @include_proc = include_proc
   end
 
   def to_s