+++ lib/graphql/execution_error.rb
@@ -2,7 +2,7 @@ module GraphQL
   # If a field's resolve function returns a {ExecutionError},
   # the error will be inserted into the response's `"errors"` key
   # and the field will resolve to `nil`.
-  class ExecutionError < RuntimeError
+  class ExecutionError < GraphQL::Error
     # @return [GraphQL::Language::Nodes::Field] the field where the error occured
     attr_accessor :ast_node
 
@@ -11,9 +11,7 @@ module GraphQL
       hash = {
         "message" => message,
       }
-      if ast_node.nil?
-        hash["locations"] = []
-      else
+      if ast_node
         hash["locations"] = [
           {
             "line" => ast_node.line,