+++ lib/graphql/input_object_type.rb
@@ -8,34 +8,59 @@
 #   end
 #
 class GraphQL::InputObjectType < GraphQL::BaseType
-  attr_accessor :name, :description, :input_fields
-  defined_by_config :name, :description, :input_fields
-  alias :arguments :input_fields
+  accepts_definitions input_field: GraphQL::Define::AssignArgument
 
-  def input_fields=(new_fields)
-    @input_fields = GraphQL::DefinitionHelpers::StringNamedHash.new(new_fields).to_h
+  # @return [Hash<String, GraphQL::Argument>] Map String argument names to their {GraphQL::Argument} implementations
+  attr_accessor :arguments
+
+  alias :input_fields :arguments
+
+  def initialize
+    @arguments = {}
   end
 
   def kind
     GraphQL::TypeKinds::INPUT_OBJECT
   end
 
-  def valid_non_null_input?(input)
-    return false unless input.is_a?(Hash) || input.is_a?(GraphQL::Query::Arguments)
-    return false unless input.all? { |name, value| input_fields[name] }
-    input_fields.all? { |name, field| field.type.valid_input?(input[name]) }
+  def validate_non_null_input(input)
+    result = GraphQL::Query::InputValidationResult.new
+
+    # Items in the input that are unexpected
+    input.each do |name, value|
+      if arguments[name].nil?
+        result.add_problem("Field is not defined on #{self.name}", [name])
+      end
+    end
+
+    # Items in the input that are expected, but have invalid values
+    invalid_fields = arguments.map do |name, field|
+      field_result = field.type.validate_input(input[name])
+      if !field_result.valid?
+        result.merge_result!(name, field_result)
+      end
+    end
+
+    result
   end
 
   def coerce_non_null_input(value)
     input_values = {}
-    input_fields.each do |input_key, input_field_defn|
+
+    arguments.each do |input_key, input_field_defn|
       field_value = value[input_key]
       field_value = input_field_defn.type.coerce_input(field_value)
+
+      # Try getting the default value
       if field_value.nil?
         field_value = input_field_defn.default_value
       end
-      input_values[input_key] = field_value unless field_value.nil?
+
+      if !field_value.nil?
+        input_values[input_key] = field_value
+      end
     end
+
     GraphQL::Query::Arguments.new(input_values)
   end
 end