+++ lib/graphql/list_type.rb
@@ -17,9 +17,19 @@ class GraphQL::ListType < GraphQL::BaseType
     "[#{of_type.to_s}]"
   end
 
-  def valid_non_null_input?(value)
-    ensure_array(value).all?{ |item| of_type.valid_input?(item) }
+  def validate_non_null_input(value)
+    result = GraphQL::Query::InputValidationResult.new
+
+    ensure_array(value).each_with_index do |item, index|
+      item_result = of_type.validate_input(item)
+      if !item_result.valid?
+        result.merge_result!(index, item_result)
+      end
+    end
+
+    result
   end
+
 
   def coerce_non_null_input(value)
     ensure_array(value).map{ |item| of_type.coerce_input(item) }