+++ lib/graphql/schema/printer.rb
@@ -48,7 +48,7 @@ module GraphQL
     module TypeKindPrinters
       module FieldPrinter
         def print_fields(type)
-          type.fields.values.map{ |field| "  #{field.name}#{print_args(field)}: #{field.type}" }.join("\n")
+          type.all_fields.map{ |field| "  #{field.name}#{print_args(field)}: #{field.type}" }.join("\n")
         end
 
         def print_args(field)
@@ -57,7 +57,12 @@ module GraphQL
         end
 
         def print_input_value(arg)
-          default_string = " = #{print_value(arg.default_value, arg.type)}" unless arg.default_value.nil?
+          if arg.default_value.nil?
+            default_string = nil
+          else
+            default_string = " = #{print_value(arg.default_value, arg.type)}"
+          end
+
           "#{arg.name}: #{arg.type.to_s}#{default_string}"
         end
 
@@ -98,7 +103,11 @@ module GraphQL
       class ObjectPrinter
         extend FieldPrinter
         def self.print(type)
-          implementations = " implements #{type.interfaces.map(&:to_s).join(", ")}" unless type.interfaces.empty?
+          if type.interfaces.any?
+            implementations = " implements #{type.interfaces.map(&:to_s).join(", ")}"
+          else
+            implementations = nil
+          end
           "type #{type.name}#{implementations} {\n#{print_fields(type)}\n}"
         end
       end