+++ spec/graphql/static_validation/rules/argument_literals_are_compatible_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
 
 describe GraphQL::StaticValidation::ArgumentLiteralsAreCompatible do
-  let(:document) { GraphQL.parse(%|
+  let(:query_string) {%|
     query getCheese {
       cheese(id: "aasdlkfj") { source }
       cheese(id: 1) { source @skip(if: {id: 1})}
@@ -9,15 +9,17 @@ describe GraphQL::StaticValidation::ArgumentLiteralsAreCompatible do
       badSource: searchDairy(product: [{source: 1.1}]) { source }
       missingSource: searchDairy(product: [{fatContent: 1.1}]) { source }
       listCoerce: cheese(id: 1) { similarCheese(source: YAK) }
+      missingInputField: searchDairy(product: [{source: YAK, wacky: 1}])
     }
 
     fragment cheeseFields on Cheese {
       similarCheese(source: 4.5)
     }
-  |)}
+  |}
 
   let(:validator) { GraphQL::StaticValidation::Validator.new(schema: DummySchema, rules: [GraphQL::StaticValidation::ArgumentLiteralsAreCompatible]) }
-  let(:errors) { validator.validate(document) }
+  let(:query) { GraphQL::Query.new(DummySchema, query_string) }
+  let(:errors) { validator.validate(query) }
 
   it 'finds undefined or missing-required arguments to fields and directives' do
     assert_equal(6, errors.length)
@@ -30,7 +32,7 @@ describe GraphQL::StaticValidation::ArgumentLiteralsAreCompatible do
 
     directive_error = {
       "message"=>"Argument 'if' on Directive 'skip' has an invalid value",
-      "locations"=>[{"line"=>4, "column"=>31}]
+      "locations"=>[{"line"=>4, "column"=>30}]
     }
     assert_includes(errors, directive_error)
 
@@ -42,7 +44,7 @@ describe GraphQL::StaticValidation::ArgumentLiteralsAreCompatible do
 
     input_object_field_error = {
       "message"=>"Argument 'source' on InputObject 'DairyProductInput' has an invalid value",
-      "locations"=>[{"line"=>6, "column"=>41}]
+      "locations"=>[{"line"=>6, "column"=>40}]
     }
     assert_includes(errors, input_object_field_error)
 
@@ -54,7 +56,7 @@ describe GraphQL::StaticValidation::ArgumentLiteralsAreCompatible do
 
     fragment_error = {
       "message"=>"Argument 'source' on Field 'similarCheese' has an invalid value",
-      "locations"=>[{"line"=>12, "column"=>7}]
+      "locations"=>[{"line"=>13, "column"=>7}]
     }
     assert_includes(errors, fragment_error)
   end